requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-config-file==0.6.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configobj==5.0.6
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
humanize==4.1.0
idna==3.3
kombu==5.2.4
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.3.1
requests==2.27.1
requests-oauthlib==1.3.1
rrss==0.0.2
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
tqdm==4.64.0
tweepy==4.9.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.1.0
wrapt==1.14.1
